fix: pass height instead of forks bitmask to populate_pooled (#190)
populate_pooled was comparing a stored fork bitmask against current
forks to determine if a pooled transaction was still current. The
height field in transaction_unconfirmed_entry was being abused to
store forks since its introduction in 2018.

Now the actual block height is stored and compared, which is the
correct semantic for determining transaction currency.

Also changes chain_state::height() and branch::height() return
types from size_t to uint32_t.

+++ src/domain/include/kth/domain/chain/chain_state.hpp
@@ -165,7 +165,7 @@ struct KD_API chain_state {
     domain::config::network network() const;
 
     [[nodiscard]]
-    size_t height() const;
+    uint32_t height() const;
 
     [[nodiscard]]
     abla::state const& abla_state() const;
